==> 1_ihs4_cleaning.Rmd <==
---
title: "ihs4_cleaning"
authors: 
- "Kevin Tang (Author)"
- "Lucia Segovia de la Revilla (Collaborator)"
date: "3/3/2020"
last-modified: "2023/03/28"
output: html_document
---

## INTRO 

1) We need to identify where the data on household consumption is stored. 

For ish4, the data on food consumption was recorded as part of the household questionnaire (datasets starting w/ HH_). Within this questionnaire, data on food consumption is recorded in the module G (Food consumption). Within the module G, we find 3 that set, each one with different level of aggregation. We are interested in g1 which is food recorded at food item level (disaggregated data). 

Hence, we need to load the file hh_mod_g1 (household questionnaire module G, 1- consumption at item level). Once it is loaded, we can check that the file is loaded correctly by checking that the observations (cases) and variables match the one reported in the worldbank website (https://microdata.worldbank.org/index.php/catalog/2936/data-dictionary).



```{r}
#--- Load libraries - reminder to always load tidyverse last

library(plyr)
library(magrittr)
library(foreign)
library(readxl)
library(dplyr)

#--- Change to show no scientific notation & round to 3 decimal places
options(scipen = 10, digits=3) 

#--- Set the plot theme to come out in black and white
theme_set(theme_bw())

#--- Read in the file

ihs4 <- read.csv(here::here("data", "hh_mod_g1.csv"))
fcode <- read.csv(here::here("data", "mwi_item_code.csv"),  fileEncoding="UTF-8-BOM")
ucode <- read.csv(here::here("data", "mwi_unit_code.csv"), fileEncoding="UTF-8-BOM")

```

## CHANGE VARIABLE NAMES AND DATA CLASS

The next step, it is to standardize the variables. This is important because we are going to be using data from different countries and the variable names will changes accross quesitonnaire. However, if we want to standardize (or harmonize) our cleaning and processing structure, it is best to standardize variables names (when possible). Within the data dictionary we can find a description of each variable. And, below is the codebook with the new variable names.

Those variables with hh_g03 refer to total consumption, hh_g04 refers to purchased by the household, hh_06 refers to food from own-production, 
hh_g07 refers to food consumed from gift (and other sources). 

Then, to identify variables that report quantity will be those that has an 'a' after the number (i.e. hh_g03a), variables reporting units will be identified by 'b' (i.e. hh_g07b, or hhg07b_label, hg07b_oth) and 'c' is for sub-unit related to photo aid guidance. 

Codebook (https://microdata.worldbank.org/index.php/catalog/2939/datafile/F192)



id =            case_id              purc_quant =    hh_g04a              gift_quant =    hh_g07a
consYN =        hh_g01               purc_unitA =    hh_g04b              gift_unitA =    hh_g07b
item_code =     hh_g02               purc_unit =     hh_g04b_label        gift_unit =     hh_g07b_label
item_oth =      hh_g01_oth           purc_unit_oth = hh_g04b_oth          gift_unit_oth = hh_g07b_oth
                                     purc_unit_size= hh_g04c              gift_unit_size= hh_g07c
                                     purc_price =    hh_g05
             
cons_quant =    hh_g03a              prod_quant =    hh_g06a
cons_unitA =    hh_g03b              prod_unitA =    hh_g06b
cons_unit =     hh_g03b_label        prod_unit =     hh_g06b_label
cons_unit_oth = hh_g03b_oth          prod_unit_oth = hh_g06b_oth
cons_unit_size= hh_g03c              prod_unit_size= hh_g06c

```{r}

#Summary info
#names(ihs4)[names(ihs4) == 'case_id'] <- 'id'
names(ihs4)[names(ihs4) == 'hh_g01'] <- 'consYN'
names(ihs4)[names(ihs4) == 'hh_g02'] <- 'item_code'
names(ihs4)[names(ihs4) == 'hh_g01_oth'] <- 'item_oth'

ihs4$case_id <- as.character(ihs4$case_id)
ihs4$consYN <- as.factor(ihs4$consYN)
ihs4$item_code <- as.factor(ihs4$item_code)
ihs4$item_oth <- as.factor(ihs4$item_oth)

#G03: Quantity and units consumed of each food 
names(ihs4)[names(ihs4) == 'hh_g03a'] <- 'cons_quant'
names(ihs4)[names(ihs4) == 'hh_g03b'] <- 'cons_unitA'
names(ihs4)[names(ihs4) == 'hh_g03b_label'] <- 'cons_unit'
names(ihs4)[names(ihs4) == 'hh_g03b_oth'] <- 'cons_unit_oth'
names(ihs4)[names(ihs4) == 'hh_g03c'] <- 'cons_unit_size'

ihs4$cons_quant <- as.numeric(ihs4$cons_quant)
ihs4$cons_unitA <- as.factor(ihs4$cons_unitA)
ihs4$cons_unit <- as.factor(ihs4$cons_unit)
ihs4$cons_unit_oth <- as.factor(ihs4$cons_unit_oth)
ihs4$cons_unit_size <- as.factor(ihs4$cons_unit_size)


#G04: Quantity and units purchased of each food 
names(ihs4)[names(ihs4) == 'hh_g04a'] <- 'purc_quant'
names(ihs4)[names(ihs4) == 'hh_g04b'] <- 'purc_unitA'
names(ihs4)[names(ihs4) == 'hh_g04b_label'] <- 'purc_unit'
names(ihs4)[names(ihs4) == 'hh_g04b_oth'] <- 'purc_unit_oth'
names(ihs4)[names(ihs4) == 'hh_g04c'] <- 'purc_unit_size'
names(ihs4)[names(ihs4) == 'hh_g05'] <- 'purc_price'

ihs4$purc_quant <- as.numeric(ihs4$purc_quant)
ihs4$purc_unitA <- as.factor(ihs4$purc_unitA)
ihs4$purc_unit <- as.factor(ihs4$purc_unit)
ihs4$purc_unit_oth <- as.factor(ihs4$purc_unit_oth)
ihs4$purc_unit_size <- as.factor(ihs4$purc_unit_size)
ihs4$purc_price <- as.numeric(ihs4$purc_price)

#G06: Quantity and units home produced of each food 
names(ihs4)[names(ihs4) == 'hh_g06a'] <- 'prod_quant'
names(ihs4)[names(ihs4) == 'hh_g06b'] <- 'prod_unitA'
names(ihs4)[names(ihs4) == 'hh_g06b_label'] <- 'prod_unit'
names(ihs4)[names(ihs4) == 'hh_g06b_oth'] <- 'prod_unit_oth'
names(ihs4)[names(ihs4) == 'hh_g06c'] <- 'prod_unit_size'

ihs4$prod_quant <- as.numeric(ihs4$prod_quant)
ihs4$prod_unitA <- as.factor(ihs4$prod_unitA)
ihs4$prod_unit <- as.factor(ihs4$prod_unit)
ihs4$prod_unit_oth <-as.factor(ihs4$prod_unit_oth)
ihs4$prod_unit_size <- as.factor(ihs4$prod_unit_size)

#G07: Quantity and units gifted/other source of each food 
names(ihs4)[names(ihs4) == 'hh_g07a'] <- 'gift_quant'
names(ihs4)[names(ihs4) == 'hh_g07b'] <- 'gift_unitA'
names(ihs4)[names(ihs4) == 'hh_g07b_label'] <- 'gift_unit'
names(ihs4)[names(ihs4) == 'hh_g07b_oth'] <- 'gift_unit_oth'
names(ihs4)[names(ihs4) == 'hh_g07c'] <- 'gift_unit_size'

ihs4$gift_quant <- as.numeric(ihs4$gift_quant)
ihs4$gift_unitA <- as.factor(ihs4$gift_unitA)
ihs4$gift_unit <- as.factor(ihs4$gift_unit)
ihs4$gift_unit_oth <- as.factor(ihs4$gift_unit_oth)
ihs4$gift_unit_size <- as.factor(ihs4$gift_unit_size)

```

## MERGE IN UNIT AND ITEM LABELS

We are merging the food code with the food description and the unit with the unit description. We are also checking that the items and codes are matching. 

```{r}
ihs4 <- merge(x=ihs4, y=fcode, by.x='item_code', by.y='code', fill=-9999, all.x = TRUE)
ihs4 <- merge(x=ihs4, y=ucode, by.x='cons_unit', by.y='code', fill=-9999, all.x = TRUE)

ihs4 <- ihs4 %>% arrange(cons_unit) %>% arrange(item_code) %>% arrange(HHID)

ihs4 <- ihs4 %>% dplyr::select(case_id:consYN, item_code, item, item_oth:cons_unitA, cons_unit, unit, cons_unit_oth:gift_unit_size)

ihs4 %>%  dplyr::count(item_code, item) %>% arrange(item_code)

ihs4 %>%  filter(item_code == "102") %>% distinct(item_code, item)

```


# CLEANING UNITS LABELLED AS "OTHER"

Food items and units are coming from a standard list. However, for some NSU, when the household is reporting the food consumed in a unit that it is not in the list will be recoded and the unit code apply for 'other' is 23.


```{r}
ihs4 %>% filter(cons_unitA==23)
ihs4 %>% filter(cons_unit_oth != "") %>% dplyr::count(cons_unit_oth) %>% arrange(desc(n)) #which units are listed as "other"
ihs4 %>% filter(cons_unitA==23) %>% dplyr::count(cons_unit_oth) %>% arrange(desc(n)) #which units are listed as "other" have not been converted to standard unit

ihs4 %>% filter(., cons_unit_oth=="HEAP") %>% dplyr::select(item, item_code, cons_quant, cons_unitA, cons_unit, cons_unit_oth) %>% arrange(item)
ihs4 %>% filter(item_code=="803") %>% dplyr::select(item, item_code, cons_quant, cons_unitA, cons_unit, cons_unit_oth) %>% arrange(desc(cons_unitA))

# Generating the new variable

ihs4$unit_code_st <- ihs4$cons_unitA

```


There are a number of item units recorded as "23" but described as existing ucodes in our codebook.

We need to change these units from "23" to their appropriate units as in this analysis any item without a standard unit will be excluded. The NSU codes for use (original + amended) are stored in the variabe `code_unit_st`.


```{r}
ihs4[ihs4$cons_unit_oth == "HEAP", "unit_code_st"] <- 10
ihs4[ihs4$cons_unit_oth == "GRAMS", "unit_code_st"] <- 18
ihs4[ihs4$cons_unit_oth == "KILOGRAM", "unit_code_st"] <- 1
ihs4[ihs4$cons_unit_oth == "PAIL", "unit_code_st"] <- 4
ihs4[ihs4$cons_unit_oth == "TUBE /SACHET", "unit_code_st"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (MEDIUM)", "unit_code_st"] <- 42
ihs4[ihs4$cons_unit_oth == "HEAP (SMALL)", "unit_code_st"] <- 10
ihs4[ihs4$cons_unit_oth == "BUNDLE", "unit_code_st"] <- 8
ihs4[ihs4$cons_unit_oth == "HEAP (LARGE)", "unit_code_st"] <- "10C"
ihs4[ihs4$cons_unit_oth == "TABLESPOON", "unit_code_st"] <- 50
ihs4[ihs4$cons_unit_oth == "5L BUCKET", "unit_code_st"] <- 26
ihs4[ihs4$cons_unit_oth == "TUBE (MEDIUM)", "unit_code_st"] <- 42
ihs4[ihs4$cons_unit_oth == "PAIL (SMALL)", "unit_code_st"] <- 4
ihs4[ihs4$cons_unit_oth == "TEASPOON", "unit_code_st"] <- 20
ihs4[ihs4$cons_unit_oth == "HEAP (MEDIUM)", "unit_code_st"] <- 10
ihs4[ihs4$cons_unit_oth == "PACKET 100 GRAM", "unit_code_st"] <- 34
ihs4[ihs4$cons_unit_oth == "PAIL (LARGE)", "unit_code_st"] <- 4
ihs4[ihs4$cons_unit_oth == "TUBE (LARGE)", "unit_code_st"] <- 43
ihs4[ihs4$cons_unit_oth == "PACKET 150 GRAM", "unit_code_st"] <- 34
ihs4[ihs4$cons_unit_oth == "PIECE (SMALL)", "unit_code_st"] <- "9A"
ihs4[ihs4$cons_unit_oth == "PIECES", "unit_code_st"] <- 9
ihs4[ihs4$cons_unit_oth == "NO. 12 PLATE", "unit_code_st"] <- 7
ihs4[ihs4$cons_unit_oth == "SATCHET (LARGE)", "unit_code_st"] <- 43
ihs4[ihs4$cons_unit_oth == "PIECE MEDIUM", "unit_code_st"] <- "9B"
ihs4[ihs4$cons_unit_oth == "PIECE", "unit_code_st"] <- 9
ihs4[ihs4$cons_unit_oth == "NO. 10 PLATE", "unit_code_st"] <- 6
ihs4[ihs4$cons_unit_oth == "5 LITRES", "unit_code_st"] <- 26
ihs4[ihs4$cons_unit_oth == "PAIL (MEDIUM)", "unit_code_st"] <- "4B"
ihs4[ihs4$cons_unit_oth == "PIECE (LARGE)", "unit_code_st"] <- "9C"
ihs4[ihs4$cons_unit_oth == "MANGO", "unit_code_st"] <- 9
ihs4[ihs4$cons_unit_oth == "PACKET (SMALL)", "unit_code_st"] <- 54
ihs4[ihs4$cons_unit_oth == "SATCHET (250 GRAM)", "unit_code_st"] <- 65
ihs4[ihs4$cons_unit_oth == "SATCHET (20 GRAM)", "unit_code_st"] <- 41
ihs4[ihs4$cons_unit_oth == "TIN", "unit_code_st"] <- 72
ihs4[ihs4$cons_unit_oth == "PIECE (MEDIUM)", "unit_code_st"] <- "9B"
ihs4[ihs4$cons_unit_oth == "PACKET 250 GRAM", "unit_code_st"] <- 65
ihs4[ihs4$cons_unit_oth == "PACKETS", "unit_code_st"] <- 60
ihs4[ihs4$cons_unit_oth == "TIN (500G)", "unit_code_st"] <- 73
ihs4[ihs4$cons_unit_oth == "SACHET", "unit_code_st"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (MEDIUM)", "unit_code_st"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (500 GRAM)", "unit_code_st"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (SMALL)", "unit_code_st"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (150 GRAM)", "unit_code_st"] <- 43
ihs4[ihs4$cons_unit_oth == "BOTTLE" & ihs4$item_code == "803", "unit_code_st"] <- 15

ihs4[ihs4$cons_unit_oth == "HEAP", "cons_unit"] <- 10
ihs4[ihs4$cons_unit_oth == "GRAMS", "cons_unit"] <- 18
ihs4[ihs4$cons_unit_oth == "KILOGRAM", "cons_unit"] <- 1
ihs4[ihs4$cons_unit_oth == "PAIL", "cons_unit"] <- 4
ihs4[ihs4$cons_unit_oth == "TUBE /SACHET", "cons_unit"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (MEDIUM)", "cons_unit"] <- 42
ihs4[ihs4$cons_unit_oth == "HEAP (SMALL)", "cons_unit"] <- 10
ihs4[ihs4$cons_unit_oth == "BUNDLE", "cons_unit"] <- 8
ihs4[ihs4$cons_unit_oth == "HEAP (LARGE)", "cons_unit"] <- 10
ihs4[ihs4$cons_unit_oth == "TABLESPOON", "cons_unit"] <- 50
ihs4[ihs4$cons_unit_oth == "5L BUCKET", "cons_unit"] <- 26
ihs4[ihs4$cons_unit_oth == "TUBE (MEDIUM)", "cons_unit"] <- 42
ihs4[ihs4$cons_unit_oth == "PAIL (SMALL)", "cons_unit"] <- 4
ihs4[ihs4$cons_unit_oth == "TEASPOON", "cons_unit"] <- 20
ihs4[ihs4$cons_unit_oth == "HEAP (MEDIUM)", "cons_unit"] <- 10
ihs4[ihs4$cons_unit_oth == "PACKET 100 GRAM", "cons_unit"] <- 34
ihs4[ihs4$cons_unit_oth == "PAIL (LARGE)", "cons_unit"] <- 4
ihs4[ihs4$cons_unit_oth == "TUBE (LARGE)", "cons_unit"] <- 43
ihs4[ihs4$cons_unit_oth == "PACKET 150 GRAM", "cons_unit"] <- 34
ihs4[ihs4$cons_unit_oth == "PIECE (SMALL)", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "PIECES", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "NO. 12 PLATE", "cons_unit"] <- 7
ihs4[ihs4$cons_unit_oth == "SATCHET (LARGE)", "cons_unit"] <- 43
ihs4[ihs4$cons_unit_oth == "PIECE MEDIUM", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "PIECE", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "NO. 10 PLATE", "cons_unit"] <- 6
ihs4[ihs4$cons_unit_oth == "5 LITRES", "cons_unit"] <- 26
ihs4[ihs4$cons_unit_oth == "PAIL (MEDIUM)", "cons_unit"] <- 4
ihs4[ihs4$cons_unit_oth == "PIECE (LARGE)", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "MANGO", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "PACKET (SMALL)", "cons_unit"] <- 54
ihs4[ihs4$cons_unit_oth == "SATCHET (250 GRAM)", "cons_unit"] <- 65
ihs4[ihs4$cons_unit_oth == "SATCHET (20 GRAM)", "cons_unit"] <- 41
ihs4[ihs4$cons_unit_oth == "TIN", "cons_unit"] <- 72
ihs4[ihs4$cons_unit_oth == "PIECE (MEDIUM)", "cons_unit"] <- 9
ihs4[ihs4$cons_unit_oth == "PACKET 250 GRAM", "cons_unit"] <- 65
ihs4[ihs4$cons_unit_oth == "PACKETS", "cons_unit"] <- 60
ihs4[ihs4$cons_unit_oth == "TIN (500G)", "cons_unit"] <- 73
ihs4[ihs4$cons_unit_oth == "SACHET", "cons_unit"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (MEDIUM)", "cons_unit"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (500 GRAM)", "cons_unit"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (SMALL)", "cons_unit"] <- 22
ihs4[ihs4$cons_unit_oth == "SATCHET (150 GRAM)", "cons_unit"] <- 43
ihs4[ihs4$cons_unit_oth == "BOTTLE" & ihs4$item_code == "803", "cons_unit"] <- 15
```

# CLEANING FOODS LABELLED AS "OTHER"

```{r, echo=FALSE}

other_foods <- fcode %>% filter(grepl("other|Specify", item, ignore.case = TRUE)) %>% pull(code)

```

Similarly, some food consumed by the household are not in that list of items, hence it can be recorded under 'other'. The codes to identify "other food items" depend on the food category and we have identified as the following `r toString(other_foods)`

Changing the way the food other are allocated to avoid overwriting items that were assigned to fixed codes. For example, there is one food that has item_oth == 'GRAIN MEAL', howeve it is also coded under "Maize ufa refined (fine flour)" (102), if we run the code without filtering out those items that were already allocated to a category, then we are changing the food code from code 102 to 101, which may have some impact in the nutrient quantities due to the content of the different flours. 

```{r}
other_foods <- fcode %>% filter(grepl("other|Specify", item, ignore.case = TRUE)) %>% pull(code)

ihs4 %>% filter(grepl("\\s", item_oth) & item_code %in% other_foods) 
ihs4 %>% filter(grepl("\\s", item_oth) & !item_code %in% other_foods) %>% distinct(item_code, item)

ihs4 %>% dplyr::count(item_code, item_oth) %>% arrange(desc(n))

ihs4 %>% filter(grepl("\\s", item_oth) & item_code %in% other_foods)  %>% dplyr::count(item_code, item_oth) %>% arrange(desc(n))
ihs4 %>% filter(grepl("\\s", item_oth) & item_code %in% other_foods)  %>% dplyr::count(item_code) %>% arrange(desc(n))

ihs4 %>% filter(grepl("BIRD", item_oth))  %>% dplyr::count(item_code, item_oth) 
ihs4 %>% filter(grepl("cassava", item, ignore.case = TRUE))  %>% dplyr::count(item_code, item) 

fcode %>% filter(code=="821")

```


We are generating a new variable for the amended item codes. We are going to name it item_code_st (as for standardised code). We are not fixing those called other that are actually other foods. As we could include them as new food item matching as done for FAO work. They are left here but commented so they are not running. 


```{r}

ihs4 <- ihs4 %>% 
  mutate(item_code_st = ifelse(!grepl("\\s", item_oth) | (grepl("\\s", item_oth) & !item_code %in% other_foods), 
           as.character(item_code),
        case_when(
# item_oth %in% c('GRAIN MEAL', 'GRANMILL', 'MAIZEMEAL') ~ 101,
item_oth=='SCONES' ~ "112", 
# item_oth %in% c('FRESHBEANS', 'GROUND AND COOKED BEANS') ~ 302,
item_oth %in% c( 'GROUND AND COOKED BEANS') ~ "310", 
item_oth=='PIGEON PEAS'~ "303", 
item_oth=='PEANUT' ~ "304",   
item_oth=='COW PEAS'~ "308", 
item_oth=='SOYA' ~ "314", 
item_oth %in% c('BONONGWE', 'CHISOSOANDBONONGWE') ~  "403", 
item_oth=='CASSAVA LEAVES' ~ "406", 
item_oth=='CUCUMBER' ~ "409",  
item_oth=='PORK'~ "506",
item_oth %in% c('BIRD', 'DUCKMEAT') ~ "509",
item_oth=='IRISH POTATOES AND FLOUR' ~ "821",
item_oth %in% c('MANDASI', 'FRIED MIXTURE OF MAIZE FLOUR/BANANA AND SODA', 'MIXTURE OF MAIZE FLOUR/BANANA/SUGAR') ~ "827", 
item_oth %in% c('BAKED AFRICAN CAKE', 'CHIKONDAMOYO') ~ "836", 
item_oth=='SMALL FISH' ~ "5023", 
item_oth=='TEA' ~ "901", 
item_oth=='SODA' ~  "907", 
item_oth=='WINE' ~ "914"))) 

```

## MISC. CLEANING FUNCTIONS

Create unique ID variable
```{r}
ihs4$measure_id <- paste0(as.character(ihs4$item_code_st),"_", as.character(ihs4$unit_code_st))
ihs4$measure_id[is.na(ihs4$cons_unit)] <- NA

ihs4 %>%  dplyr::count(item_code_st, item) %>% arrange(item_code_st)

```

Saving the output 

```{r}

# Cleaned food items others & NSU recoded
saveRDS(ihs4, here::here("data","inter-output", "hh_mod_g_clean.RDS"))

```
