==> 3_nep.Rmd <==
---
title: "nonedible.portions.fctmatch"
authors: 
 - "Kevin Tang (Author)"
 - "Lucia Segovia de la Revilla (Collaborator)"
date: "11/12/2020"
udate: 29/07/2024"
output: html_document
editor_options: 
  chunk_output_type: inline
---

## Loading Packages & data

```{r}
# Loading the packages
library(dplyr)

# Loading the data
ihs4 <- read.csv(here::here("data", "inter-output", "hh_mod_g_processed.csv"))

#Edible portion
nep <- read.csv(here::here("data", "factors", "nep.factors.csv"))

# foodgroup <- read.csv("data/nutrient.supply/foodgroup.csv")

# Checking missing values
sum(ihs4$consYN == 1 & is.na(ihs4$kg_d))

```


## Edible portions of food items

This step requires us to eliminate food weight that would have been recalled as part of the HCES questionnaire, but would not have actually been consumed due to processing or preparation. This includes parts of food like banana peels, skins of fruits and tubers that would have been peeled, bones of large fish that would have been discarded etc. 

For these data, we conducted a similar matching process that done for food composition matching, except we used "non-edible food portions". These values are not available in MAFOODS, KENFCT, or LSOFCT so we opted for the values provided in the FAO West African FCTs. Missing values were filled in with assumptions and should be cross referenced to the USDA Food Composition Database. 

```{r}
#nep <- nep %>% select(item_code, nep)
nep$nep <-as.numeric(nep$nep)

# Merging edible portion (EP) and consumption
ihs4 <- merge(x=ihs4, y=nep , by.x='item_code', by.y='item_code', fill=-9999, all.x = TRUE) %>% arrange(HHID)

# Applying the EP
ihs4$kg_d_nep <- ihs4$kg_d * ihs4$nep
# Changing units to g/day per EP
ihs4$g_d_nep <-ihs4$kg_d_nep*1000
```

## Labeling food groups of food items

We need to add a variable to group items by the food group that a food item fell under according to the IHS. There are 11 food groups in total: 
-Cereals/grains
-Roots/tubers/plantains
-Nuts/pulses
-Vegetables
-Meat/animal products
-Fruit
-Dairy
-Fats
-Spices/misc
-Vendor food
-Beverages 

```{r}
ihs4 <- merge(x=ihs4, y=foodgroup, by.x='item_code', by.y='code', fill=-9999, all.x = TRUE) %>% arrange(item_code) %>% arrange(HHID)
```

## Saving the file

```{r, eval=FALSE}

write.csv(ihs4, here::here("data", "inter-output", "hh_mod_g_nep.csv"))

```
